+++ src/profile/jwks.test.ts
@@ -20,7 +20,7 @@
 //
 
 import { describe, it, jest, expect } from "@jest/globals";
-import { Response } from "cross-fetch";
+import { Response } from "@inrupt/universal-fetch";
 import { rdf, security } from "../constants";
 import { mockSolidDatasetFrom } from "../resource/mock";
 import { buildThing } from "../thing/build";