+++ src/resource/resource.test.ts
@@ -21,7 +21,7 @@
 
 import { jest, describe, it, expect } from "@jest/globals";
 
-import { Response } from "cross-fetch";
+import { Response } from "@inrupt/universal-fetch";
 import {
   getResourceInfo,
   getSourceIri,
@@ -42,6 +42,7 @@ import {
   SolidClientError,
 } from "../interfaces";
 import { createSolidDataset } from "./solidDataset";
+import { mockResponse } from "../tests.internal";
 
 jest.mock("../fetcher.ts", () => ({
   fetch: jest.fn().mockImplementation(() =>
@@ -52,13 +53,6 @@ jest.mock("../fetcher.ts", () => ({
     )
   ),
 }));
-
-function mockResponse(
-  body?: BodyInit | null,
-  init?: ResponseInit & { url: string }
-): Response {
-  return new Response(body, init);
-}
 
 describe("getResourceInfo", () => {
   it("calls the included fetcher by default", async () => {
@@ -90,10 +84,8 @@ describe("getResourceInfo", () => {
   it("keeps track of where the SolidDataset was fetched from", async () => {
     const mockFetch = jest
       .fn<typeof fetch>()
-      .mockReturnValue(
-        Promise.resolve(
-          mockResponse(undefined, { url: "https://some.pod/resource" })
-        )
+      .mockResolvedValue(
+        mockResponse(undefined, undefined, "https://some.pod/resource")
       );
 
     const solidDatasetInfo = await getResourceInfo(
@@ -111,10 +103,13 @@ describe("getResourceInfo", () => {
   it("knows when the Resource contains a SolidDataset", async () => {
     const mockFetch = jest.fn<typeof fetch>().mockReturnValue(
       Promise.resolve(
-        mockResponse(undefined, {
-          url: "https://arbitrary.pod/resource",
-          headers: { "Content-Type": "text/turtle" },
-        })
+        mockResponse(
+          undefined,
+          {
+            headers: { "Content-Type": "text/turtle" },
+          },
+          "https://arbitrary.pod/resource"
+        )
       )
     );
 
@@ -131,10 +126,13 @@ describe("getResourceInfo", () => {
   it("knows when the Resource does not contain a SolidDataset", async () => {
     const mockFetch = jest.fn<typeof fetch>().mockReturnValue(
       Promise.resolve(
-        mockResponse(undefined, {
-          url: "https://arbitrary.pod/resource",
-          headers: { "Content-Type": "image/svg+xml" },
-        })
+        mockResponse(
+          undefined,
+          {
+            headers: { "Content-Type": "image/svg+xml" },
+          },
+          "https://arbitrary.pod/resource"
+        )
       )
     );
 
@@ -153,7 +151,7 @@ describe("getResourceInfo", () => {
       .fn<typeof fetch>()
       .mockReturnValue(
         Promise.resolve(
-          mockResponse(undefined, { url: "https://arbitrary.pod/resource" })
+          mockResponse(undefined, undefined, "https://arbitrary.pod/resource")
         )
       );
 
@@ -170,10 +168,13 @@ describe("getResourceInfo", () => {
   it("exposes the Content Type when known", async () => {
     const mockFetch = jest.fn<typeof fetch>().mockReturnValue(
       Promise.resolve(
-        mockResponse(undefined, {
-          url: "https://some.pod/resource",
-          headers: { "Content-Type": "text/turtle; charset=UTF-8" },
-        })
+        mockResponse(
+          undefined,
+          {
+            headers: { "Content-Type": "text/turtle; charset=UTF-8" },
+          },
+          "https://some.pod/resource"
+        )
       )
     );
 
@@ -207,12 +208,15 @@ describe("getResourceInfo", () => {
   it("provides the IRI of the relevant ACL resource, if provided", async () => {
     const mockFetch = jest.fn<typeof fetch>().mockReturnValue(
       Promise.resolve(
-        mockResponse(undefined, {
-          headers: {
-            Link: '<aclresource.acl>; rel="acl"',
+        mockResponse(
+          undefined,
+          {
+            headers: {
+              Link: '<aclresource.acl>; rel="acl"',
+            },
           },
-          url: "https://some.pod/container/resource",
-        })
+          "https://some.pod/container/resource"
+        )
       )
     );
 
@@ -229,12 +233,15 @@ describe("getResourceInfo", () => {
   it("exposes the URLs of linked Resources", async () => {
     const mockFetch = jest.fn<typeof fetch>().mockReturnValue(
       Promise.resolve(
-        mockResponse(undefined, {
-          headers: {
-            Link: '<aclresource.acl>; rel="acl", <https://some.pod/profile#WebId>; rel="http://www.w3.org/ns/solid/terms#podOwner", <https://some.pod/rss>; rel="alternate", <https://some.pod/atom>; rel="alternate"',
+        mockResponse(
+          undefined,
+          {
+            headers: {
+              Link: '<aclresource.acl>; rel="acl", <https://some.pod/profile#WebId>; rel="http://www.w3.org/ns/solid/terms#podOwner", <https://some.pod/rss>; rel="alternate", <https://some.pod/atom>; rel="alternate"',
+            },
           },
-          url: "https://some.pod",
-        })
+          "https://some.pod"
+        )
       )
     );
 
@@ -253,11 +260,9 @@ describe("getResourceInfo", () => {
   });
 
   it("exposes when no Resources were linked", async () => {
-    const mockFetch = jest.fn<typeof fetch>().mockResolvedValue(
-      mockResponse(undefined, {
-        url: "https://arbitrary.pod",
-      })
-    );
+    const mockFetch = jest
+      .fn<typeof fetch>()
+      .mockResolvedValue(mockResponse(undefined, {}, "https://arbitrary.pod"));
 
     const solidDatasetInfo = await getResourceInfo(
       "https://some.pod/container/resource",
@@ -268,15 +273,16 @@ describe("getResourceInfo", () => {
   });
 
   it("does not provide an IRI to an ACL resource if not provided one by the server", async () => {
-    const mockResponse = new Response(undefined, {
-      headers: {
-        Link: '<arbitrary-resource>; rel="not-acl"',
+    const mockedResponse = mockResponse(
+      undefined,
+      {
+        headers: {
+          Link: '<arbitrary-resource>; rel="not-acl"',
+        },
       },
-      url: "https://arbitrary.pod",
-      // We need the type assertion because in non-mock situations,
-      // you cannot set the URL manually:
-    } as ResponseInit);
-    const mockFetch = jest.fn<typeof fetch>().mockResolvedValue(mockResponse);
+      "https://arbitrary.pod"
+    );
+    const mockFetch = jest.fn<typeof fetch>().mockResolvedValue(mockedResponse);
 
     const solidDatasetInfo = await getResourceInfo(
       "https://some.pod/container/resource",
@@ -373,7 +379,7 @@ describe("getResourceInfo", () => {
       .fn<typeof fetch>()
       .mockReturnValue(
         Promise.resolve(
-          mockResponse(undefined, { url: "https://some.pod/resource" })
+          mockResponse(undefined, undefined, "https://some.pod/resource")
         )
       );
 
@@ -387,7 +393,10 @@ describe("getResourceInfo", () => {
   });
 
   it("returns a meaningful error when the server returns a 403", async () => {
-    const mockResponse = new Response("Not allowed", { status: 403 });
+    const mockResponse = new Response("Not allowed", {
+      status: 403,
+      statusText: "Forbidden",
+    });
     jest
       .spyOn(mockResponse, "url", "get")
       .mockReturnValue("https://some.pod/resource");
@@ -405,13 +414,18 @@ describe("getResourceInfo", () => {
   });
 
   it("overrides a 403 error if provided the appropriate option", async () => {
-    const mockFetch = jest
-      .fn<typeof fetch>()
-      .mockReturnValue(
-        Promise.resolve(
-          mockResponse("Forbidden", { status: 403, url: "https://some.url" })
+    const mockFetch = jest.fn<typeof fetch>().mockReturnValue(
+      Promise.resolve(
+        mockResponse(
+          "Forbidden",
+          {
+            status: 403,
+            statusText: "Forbidden",
+          },
+          "https://some.url"
         )
-      );
+      )
+    );
 
     const resourceInfo = await getResourceInfo("https://some.pod/resource", {
       fetch: mockFetch,
@@ -424,13 +438,16 @@ describe("getResourceInfo", () => {
   });
 
   it("returns a meaningful error when the server returns a 404", async () => {
-    const mockResponse = new Response("Not found", { status: 404 });
+    const mockedResponse = new Response("Not found", {
+      status: 404,
+      statusText: "Not Found",
+    });
     jest
-      .spyOn(mockResponse, "url", "get")
+      .spyOn(mockedResponse, "url", "get")
       .mockReturnValue("https://some.pod/resource");
     const mockFetch = jest
       .fn<typeof fetch>()
-      .mockReturnValue(Promise.resolve(mockResponse));
+      .mockReturnValue(Promise.resolve(mockedResponse));
 
     const fetchPromise = getResourceInfo("https://some.pod/resource", {
       fetch: mockFetch,